backend/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
datefinder==0.6.1
Django==2.0.4
django-cors-headers==2.2.0
django-vote==2.1.7
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
lxml==4.2.1
MarkupSafe==1.0
nltk==3.2.5
numpy==1.14.3
pandas==0.22.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyshorteners==0.6.1
python-dateutil==2.7.2
pytz==2018.4
regex==2016.1.10
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
uritemplate==3.0.0
urllib3==1.22
